lean-core: prune final 5 plugins → maw-js is 7 INFRA only
Vendors helpers from the 4 deferred plugins (bud/pair/overview/soul-sync)
plus cross-team-queue (Phase 0 leftover) into src/lib/ + src/shared/, then
prunes the dirs.

- pair/codes.ts → src/lib/pair-codes.ts (consumed by api/pair + core/consent/pin)
- overview/impl.processMirror → src/lib/process-mirror.ts (consumed by api/sessions)
- cross-team-queue/types → src/shared/cross-team-queue.types.ts (folded inline; was a re-export)
- bud SDK re-exports DROPPED (community plugin → not in SDK surface)
- soul-sync/bud — only inbound ref was bud→soul-sync (now both pruned together)

Stranded test files removed: 9 test/{bud,pair,soul-sync}-*.test.ts —
plugin tests live in the registry now.

End state: src/commands/plugins/ has 7 INFRA plugins
(federation, fleet, oracle, plugin, session, tmux, transport).

Build: 624 modules / 0.85 MB (down from 640 / 0.88 MB).
test:isolated: 4 baseline failures (fleet-doctor, oracle-manifest,
resolve-local-first, resolve-psi) — same as alpha tip.

Related: #918, #919, #921, #922, #923, #926. Closes lean-core extraction.
