--- src/commands/plugins/bud/plugin.json
@@ -1,34 +1,0 @@
-{
-  "name": "bud",
-  "version": "1.0.0",
-  "entry": "./index.ts",
-  "sdk": "^1.0.0",
-  "description": "Create a new oracle (bud from parent)",
-  "cli": {
-    "command": "bud",
-    "help": "maw bud <name> [--from <oracle>] [--root] [--seed] [--org <org>] [--repo org/repo] [--issue N] [--note <text>] [--fast] [--split] [--dry-run]\n       Or:    maw bud --from-repo <path> --stem <stem> [--pr] [--dry-run]  (#588, scaffold-only)\n       Default: born blank. Use --seed to pre-load parent's ψ at birth.\n       Pull memory later: maw soul-sync <parent> --from\n       NOTE: <name> is the STEM. Repo created as <name>-oracle.\n       e.g. 'maw bud fusion' → 'fusion-oracle'. 'maw bud fusion-oracle' → 'fusion-oracle-oracle' ✗",
-    "flags": {
-      "--from": "string",
-      "--from-repo": "string",
-      "--stem": "string",
-      "--org": "string",
-      "--repo": "string",
-      "--issue": "number",
-      "--note": "string",
-      "--fast": "boolean",
-      "--root": "boolean",
-      "--blank": "boolean",
-      "--pr": "boolean",
-      "--split": "boolean",
-      "--seed": "boolean",
-      "--dry-run": "boolean"
-    }
-  },
-  "api": {
-    "path": "/api/bud",
-    "methods": [
-      "POST"
-    ]
-  },
-  "weight": 0
-}
