--- src/commands/plugins/cross-team-queue/plugin.json
@@ -1,18 +1,0 @@
-{
-  "name": "cross-team-queue",
-  "version": "1.0.0",
-  "entry": "./index.ts",
-  "sdk": "^1.0.0",
-  "description": "Unified inbox view across oracle vaults — scans ψ/memory/<oracle>/inbox/*.md, parses frontmatter, returns {items, stats, errors}.",
-  "author": "Soul-Brews-Studio",
-  "cli": {
-    "command": "cross-team-queue",
-    "aliases": ["ctq"],
-    "help": "maw cross-team-queue [--recipient=] [--team=] [--type=] [--max-age-hours=] — unified inbox across oracle vaults"
-  },
-  "api": {
-    "path": "/api/plugins/cross-team-queue",
-    "methods": ["GET"]
-  },
-  "weight": 50
-}
