--- src/commands/plugins/overview/index.ts
@@ -1,31 +1,0 @@
-import type { InvokeContext, InvokeResult } from "../../../plugin/types";
-import { cmdOverview } from "./impl";
-
-export const command = {
-  name: "overview",
-  description: "Show fleet overview / war room dashboard.",
-};
-
-export default async function handler(ctx: InvokeContext): Promise<InvokeResult> {
-  const logs: string[] = [];
-  const origLog = console.log;
-  const origError = console.error;
-  console.log = (...a: any[]) => {
-    if (ctx.writer) ctx.writer(...a);
-    else logs.push(a.map(String).join(" "));
-  };
-  console.error = (...a: any[]) => {
-    if (ctx.writer) ctx.writer(...a);
-    else logs.push(a.map(String).join(" "));
-  };
-  try {
-    const args = ctx.source === "cli" ? (ctx.args as string[]) : [];
-    await cmdOverview(args);
-    return { ok: true, output: logs.join("\n") || undefined };
-  } catch (e: any) {
-    return { ok: false, error: logs.join("\n") || e.message, output: logs.join("\n") || undefined };
-  } finally {
-    console.log = origLog;
-    console.error = origError;
-  }
-}
