--- src/commands/plugins/overview/plugin.json
@@ -1,17 +1,0 @@
-{
-  "name": "overview",
-  "version": "1.0.0",
-  "entry": "./index.ts",
-  "sdk": "^1.0.0",
-  "description": "Show fleet overview / war room dashboard.",
-  "author": "Soul-Brews-Studio",
-  "cli": {
-    "command": "overview",
-    "aliases": [
-      "warroom",
-      "ov"
-    ],
-    "help": "maw overview [args] — show fleet overview / war room dashboard"
-  },
-  "weight": 10
-}
