+++ src/lib/pair-codes.ts
